# model/entity/participant.go
package entity

// Participant example
type Participant struct {
	Id   int64  `json:"id"`
	Name string `json:"name"`
}

// ParticipantInfo example
type ParticipantInfo struct {
	Name string `json:"name"`
}